+++ frontend-streamhub-main/src/main.tsx
@@ -8,33 +8,7 @@ import HomePage from "@/pages/HomePage";
 import SeriesPage from "@/pages/SeriesPage";
 import WatchPage from "@/pages/WatchPage";
 import WatchPartyPage from "@/pages/WatchPartyPage";
-
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Root />,
-    errorElement: <ErrorPage />,
-    children: [
-      {
-        index: true,
-        element: <HomePage />,
-      },
-      {
-        path: "watch/:seriesId",
-        element: <SeriesPage />,
-      },
-      {
-        path: "watch/:seriesId/:episodeId",
-        element: <WatchPage />,
-      },
-      {
-        path: "watch-party/:sessionId",
-        element: <WatchPartyPage />,
-      },
-    ],
-  },
-]);
+import RootRun from "@/pages/RootRun.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
