+++ frontend-streamhub-main/src/routes/root.tsx
@@ -1,12 +1,8 @@
 import { Outlet } from "react-router-dom";
+import Navbar from "@/pages/Navbar.tsx";
+import React from "react";
 
-const Navbar = () => {
-    return (
-        <div className="h-[12.5vh] bg-black flex justify-center items-center">
-            <p className="text-white text-2xl">Navbar Mock</p>
-        </div>
-    )
-}
+
 
 const Footer = () => {
     return (
@@ -19,7 +15,7 @@ const Footer = () => {
 const Root = () => {
     return (
         <div className="bg-gradient-to-b from-indigo-900 to-neutral-900">
-            <Navbar />
+            <Navbar></Navbar>
             <div className="min-h-[80vh] mx-20 my-4 border-2">
                 <Outlet />
             </div>
