+++ frontend-streamhub-main/tsconfig.app.json
@@ -30,7 +30,8 @@
       "@/*": [
         "./src/*"
       ]
-    }
+    },
+    "allowSyntheticDefaultImports": true
   },
   "include": ["src"]
 }
